+++ CHANGELOG.md
@@ -11,9 +11,17 @@
 
 - Set page link aria label function with `page_link_aria_label_fun` attribute
   on `Flop.Phoenix.pagination/1` component instead of using `opts`.
-- Set pagination list attributes with `list_attrs` attribute instead of
+- Set pagination list attributes with `page_list_attrs` attribute instead of
   `pagination_list_attrs` option.
-- Remove default class for pagination list.
+- Set pagination list item attributes with `page_list_item_attrs` attribute
+  instead of `pagination_list_item_attrs` option.
+- Set pagination link attributes with `page_link_attrs` attribute
+  instead of `pagination_link_attrs` option.
+- Set current pagination link attributes with `current_page_link_attrs`
+  attribute instead of `current_link_attrs` option.
+- Set disabled link attributes with `disabled_link_attrs` attribute instead of
+  `disabled_attrs` option.
+- Remove default class for pagination list and pagination links.
 - Use `<button>` elements for pagination if no `path` is set.
 - Add `rel` attribute to previous/next links.
 - Mark up disabled previous/next links of the pagination as
@@ -24,6 +32,8 @@
 
 ### Removed
 
+- Remove `opts` attribute and configuration via application environment from
+  pagination component.
 - Remove `wrapper_attrs` from the pagination options. Pass the attributes
   directly as attributes instead.
 - Remove default `class` attributes from pagination component.
@@ -34,10 +44,6 @@
   `next` slot instead.
 - Remove `ellipsis_attrs` and `ellipsis_content` from the pagination options.
   Use the `ellipsis` slot instead.
-- Remove `current_link_attrs` from pagination options. Use
-  `[aria-current="page"]` CSS selector instead.
-- Remove `disabled_attrs` from pagination options. Use `[aria-disabled="true"]`
-  CSS selector instead.
 
 ### How to upgrade
 
@@ -108,7 +114,8 @@ Replace the `:ellipsis_attrs` and `:ellipsis_content` attributes with the
 ```
 
 Remove the `:disabled_attrs` option. Select disabled links in CSS with
-`a:[aria-disabled="true"]`.
+`a:[aria-disabled="true"]`. Note that the `disabled` and `aria-disabled`
+attributes are set automatically and do not need to be passed here.
 
 ```diff
   <Flop.Phoenix.pagination
@@ -117,6 +124,7 @@ Remove the `:disabled_attrs` option. Select disabled links in CSS with
     opts={[
 -     disabled_attrs: [class: "is-disabled"]
     ]}
++     disabled_link_attrs: [class: "is-disabled"]
   >
 ```
 
@@ -134,8 +142,12 @@ Remove the `:pagination_link_aria_label` option and set the
   >
 ```
 
-Remove the `:pagination_list_attrs` option and set the `list_attrs` attribute
-instead. If you relied on the default class, set it explicitly.
+Remove the `:pagination_list_attrs`, `:pagination_list_item_attrs`,
+`:pagination_link_attrs`, and `:current_link_attrs` options and set the
+`page_list_attrs`, `page_list_item_attrs`, `page_link_attrs`, and
+`current_page_link_attrs` attributes instead. If you relied on the default page
+list class and page link classes, set them explicitly. Note that the
+`aria-current` attribute is set automatically and does not need to be set here.
 
 ```diff
   <Flop.Phoenix.pagination
@@ -143,8 +155,17 @@ instead. If you relied on the default class, set it explicitly.
     path={@path}
     opts={[
 -     pagination_list_attrs: [class: "pagination-list"]
+-     pagination_list_item_attrs: [class: "pagination-list-item"]
+-     pagination_link_attrs: [class: "pagination-link"]
+-     current_link_attrs: [
+-       class: "pagination-link is-current",
+-       aria: [current: "page"]
+-     ]
     ]}
-+     list_attrs={[class: "pagination-list"]}
++     page_list_attrs={[class: "pagination-list"]}
++     page_list_item_attrs={[class: "pagination-list-item"]}
++     page_link_attrs={[class: "pagination-link"]}
++     current_page_link_attrs={[class: "pagination-link is-current"]}
   >
 ```